Add status update messages to Rviz plugin about user commands and robot actions.
Informs the user about what the robot is currently doing and any of the commands the robot may have received from the user. The updates can be seen at the top of the Rviz plugin box in Rviz. 

The status currently updates each time the robot receives a command or performs a task and doesn't change until the next status update comes in. This is unhelpful when the robot receives more than one of the same command in a row and then it can be difficult to tell from the status updates what the robot is currently doing. 

Future work might include figuring out how long each status update should appear for.